requirements.txt
werkzeug
flask
flask-restful
pdf2image
pillow
pyqrcode
pyzbar
pdf2image
